Move editor related utilities into `src/utils/editor` and update imports.
