+++ src/actions/navigation.js
@@ -1,6 +1,6 @@
 const constants = require("../constants");
 const { clearSourceMaps } = require("../utils/source-map");
-const { clearDocuments } = require("../utils/source-documents");
+const { clearDocuments } = require("../utils/editor/source-documents");
 const { getSources } = require("../reducers/sources");
 const { waitForMs } = require("../utils/utils");
 const { newSources } = require("./sources");
