+++ src/utils/editor/source-search.js
